# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ShuKnow API
  description: REST API for the ShuKnow AI-powered file organization platform.
  version: 1.0.0
  contact:
    name: ShuKnow Team
servers:
  - url: https://localhost:5432
    description: Local development
...
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /api/files/{fileId}:
    parameters:
      - $ref: '#/components/parameters/FileId'

    get:
      operationId: getFile
      tags: [Files]
      summary: Get file metadata
      description: Returns file metadata including name, description, content type, and size.
      responses:
        '200':
          description: File metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateFile
      tags: [Files]
      summary: Update file name and/or description
      description: Validates name uniqueness within the file's current folder.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFileRequest'
      responses:
        '200':
          description: File metadata updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: A file with this name already exists in the folder.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    delete:
      operationId: deleteFile
      tags: [Files]
      summary: Permanently delete a file
      responses:
        '204':
          description: File deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/files/{fileId}/content:
    parameters:
      - $ref: '#/components/parameters/FileId'

    get:
      operationId: downloadFileContent
      tags: [Files]
      summary: Download file content
      description: |
        Streams the actual file content. Returns the stored MIME type or
        `application/octet-stream` as fallback. Supports the `Range` header
        for partial content / large file downloads.
      parameters:
        - name: Range
          in: header
          required: false
          schema:
            type: string
          description: 'HTTP Range header for partial downloads (e.g. `bytes=0-1023`).'
      responses:
        '200':
          description: Full file content.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '206':
          description: Partial content (range request satisfied).
          headers:
            Content-Range:
              schema:
                type: string
                example: bytes 0-1023/4096
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '416':
          description: Range not satisfiable.

    put:
      operationId: replaceFileContent
      tags: [Files]
      summary: Replace file content
      description: |
        Uploads new content to replace the existing file. The file's metadata
        (name, description) is unchanged.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Content replaced. Returns updated file metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          description: File exceeds the maximum allowed size.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

    patch:
      operationId: updateTextContent
      tags: [Files]
      summary: Update text content of a text-based file
      description: |
        Lightweight text content update for text-based files. Only allowed when
        the file's contentType starts with text/. Returns 409 Conflict if the
        file is not text-based.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTextContentRequest'
      responses:
        '200':
          description: Content updated. Returns updated file metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: File is not text-based.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/files/{fileId}/move:
    parameters:
      - $ref: '#/components/parameters/FileId'

    patch:
      operationId: moveFile
      tags: [Files]
      summary: Move a file to a different folder
      description: Validates name uniqueness in the target folder.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveFileRequest'
      responses:
        '200':
          description: File moved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDto'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: A file with this name already exists in the target folder.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/files/{fileId}/reorder:
    parameters:
      - $ref: '#/components/parameters/FileId'

    patch:
      operationId: reorderFile
      tags: [Files]
      summary: Reorder a file within its folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderFileRequest'
      responses:
        '204':
          description: File reordered.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ──────────────────────────────────────────────
  # Chat
  # ──────────────────────────────────────────────
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token in the `Authorization` header.

    CookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token in an HTTP-only cookie (`SameSite=None; Secure`).

  # ──────────────────────────────────────────────
  # Reusable Parameters
  # ──────────────────────────────────────────────
  parameters:
...
    FileId:
      name: fileId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the file.
...
  responses:
    Unauthorized:
      description: Authentication required. JWT is missing, invalid, or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ValidationError:
      description: One or more validation errors occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationProblemDetails'

  # ──────────────────────────────────────────────
  # Schemas
  # ──────────────────────────────────────────────
  schemas:
...
    UpdateFileRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
          maxLength: 1000
    MoveFileRequest:
      type: object
      required: [targetFolderId]
      properties:
        targetFolderId:
          type: string
          format: uuid
          description: Destination folder ID.
    UpdateTextContentRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          description: New text content for the file.
    ReorderFileRequest:
      type: object
      required: [position]
      properties:
        position:
          type: integer
          minimum: 0
          description: 0-based target index among siblings.
          example: 2
    FileDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        folderId:
          type: string
          format: uuid
        folderName:
          type: string
          example: Invoices
        name:
          type: string
          example: invoice-2024-01.pdf
        description:
          type: string
          example: January 2024 electricity bill
        contentType:
          type: string
          example: application/pdf
        sizeBytes:
          type: integer
          format: int64
          example: 204800
        sortOrder:
          type: integer
          description: Position within the parent folder.
          example: 0
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp.

    # ── Chat ──────────────────────────────────
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details.
      properties:
        type:
          type: string
          example: https://tools.ietf.org/html/rfc7807
        title:
          type: string
          example: Not Found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: Object with ID '...' was not found.
        instance:
          type: string
    ValidationProblemDetails:
      allOf:
        - $ref: '#/components/schemas/ProblemDetails'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
              example:
                name: ["The Name field is required."]
                description: ["Description must not exceed 1000 characters."]